requirements.txt
torch==1.13.1+cu117
pyyaml
loguru
tqdm
tensorboard